requirements.txt
streamlit==1.26.0
pandas==2.0.3
numpy==1.24.3
plotly==5.15.0
matplotlib==3.7.2
textblob==0.17.1
nltk==3.8.1
python-dotenv==1.0.0
langchain-groq==0.1.0
langchain-core==0.1.0
fpdf2==2.7.5
